State.Channels.updateSidebar: take an optional TeamId argument
This change makes it so that updateSidebar can be used to update the
sidebar for a specific team (if given a Just team ID) or all teams (if
given Nothing). This is necessary because some application state changes
(such as whether certain DM channels are visible) need to affect all
sidebars, whereas other changes (such as whether we are a member of a
specific channel in a team) need to affect only the sidebar for that
team.