Types: remove appMode, add csTeam(tId) lens, add csTeams field
This change replaces the fixed TeamState in the ChatState
(csCurrentTeam) with a map from team ID to TeamState (csTeams). It also
modifies csCurrentTeam to just look up in that map based on the
csCurrentTeamId value.

This change also removed the appMode function, which always assumed the
caller wanted the mode of the current team, and forced all call sites to
access the mode of the desired team.

This change also extended the server base URL functions to require a
team ID since they assumed the current team before.