ShowHelp: improve formatting of generated docs
This change uses a more appropriate data type (instead of Either) to
indicate which bits of text should be verbatim-formatted vs. plain. It
also uses that data type to format event name strings so we can avoid
verbatim formatting when it's undesirable (when indicating that an event
is unbound or when it is not bindable).