Make channel list take up full window height
This change makes it so that the channel list takes up the full window
height rather than sharing it with the message preview and editor. This
makes the editor line up with the channel messages and makes the editor
width match the channel message area width, making for more consistent
presentation of the message preview and the final posted message.