State.Teams: scroll the team the list to the left on team changes
This change ensures that the team list viewport is not left in an
arbitrary state when team changes occur. Without this change, the
contents of the team list may be obscured on the left or the right,
depending on the team change, because the viewport state is driven by
making the current team visible. We don't give the user direct control
over the team list viewport, so when it gets into that weird state,
there isn't a clean way to get out of it other than cycling through the
team list. This change makes it so that we always bias the team list in
favor of showing the leftmost (initial) list entries while still
ensuring that the currently selected team is visible. This change is
modeled after the exact same concern and behavior that biases the
channel list to show its topmost entries.