Add URL list keybinding to prompt for attachment save path
This change introduces a new key event, 'save-attachment' (default
binding: 's'), for the URL list user interface. When triggered, if the
selected URL entry is for an attachment, a dialog box is shown prompting
the user for a path to which to save the attachment. (If the entry is
not an attachment, the event is ignored.)