URL selection: display available actions in UI
This change adds a bottom bar display to the URL list to indicate any
actions that are available for the selected link, similarly to how we
display such actions for messages in message selection mode.

This change also adds a new theme attribute, 'urlSelectStatus', that is
used to render the keybindings shown in the options list.