Channel list: migrate unread flag into each entry
Rather than computing channel list entry unread status on the fly (at
drawing time and at zipper traversal time) we compute it and store it in
each channel list entry. This has the knock-on benefit that the
"hasUnread" function no longer needs to be exposed by Types.hs. This
change also removed Matterhorn.Zipper.updateList, since it is no longer
safe to compare zipper entries for equality. This is because the notion
of "equality" for channel list entries (for the purposes of preserving
selection between zipper updates) should be based solely on the channel
ID of each entry, *not* the additional metadata that now lives in those
entries such as unread status.