ChannelListEntry: carry sorting value
This change factors sorting out of the process of constructing the
channel list entries. This is done partly to generalize how sorting is
performed, but more importantly, it is done so that sorting can be
delayed until after the channel list entries are obtained. This change
is groundwork for an upcoming change to fix how the favorite channel
list entries are computed; those entries need to be obtained by
filtering the other entry lists, and then they need to be sorted by
name. This change makes it possible to do that sorting at the right
time.