Draw.Messages.renderLastMessages: refactor to preserve rendering metdata
This change refactors renderLastMessages so that instead of just
assembling Vty images and throwing away the "Result" values from Brick
(as it did before), this preserves those Results and puts them in
a vBox. This change is crucical to ensure that "clickable" regions
(extents) in any of the messages get preserved and converted into global
coordinates by vBox. Prior to this change, since that metadata was being
discarded, any mouse clicks in the message areas could never be detected
because the extents were not available in the final Result of the call
to renderLastMessages.