RichText.Flatten: avoid re-use of name sequence numbers
This change mostly updates the ENonBreaking case of 'flatten' so that
when it recurses into flattenInlineSeq', it does not throw away any
sequence number increments that happened in the recursive call. Instead
it captures the sequence number after that call and updates the monad
state to carry the sequence forward from that point. This entailed
exposing the sequence number's final value in the return value of
flattenInlineSeq'.