Events: only respond to initial mouse clicks, not drags
This change makes it so that mouse "down" events which get generated by
mouse dragging are ignored in all cases. This is because all of our
current mouse features should respond only to one click, but dragging on
a UI element generates multiple "mouse down" events and that is
undesirable.

This change works by tracking the last click event in the application
state resetting it on mouse up events or when a new mouse down event
arrives on a different UI element (in the case of a mouse up event never
arriving because the cursor left the window before the user released the
button).