renderMessageSeq: let caller customize message cropping behavior
Previously this function used vLimit to constrain individual message
renderings. In the context of message selection mode, where this is used
to render the sequences of messages before and after the selected
message, vLimit is only the right constraint to use for messages after
the selection because vLimit will crop any oversized messages from the
bottom. Because this was using vLimit for the messages preceding the
selection, that meant the topmost message would get cropped from its
bottom side, resulting in a broken message. Instead that message should
get cropped from its top side. This change makes it possible for the
caller to provide the function to the cropping; consequently we use
cropTopBy in that case.