Matterhorn.FilePaths: prefer emoji.json obtained via getDataFileName
Use cabal's internal mechanism to try and obtain the path to emoji.json.
If that path exists, we prefer it over the default
$BINPATH/../emoji/emoji.json that is currently used for release builds.

This change means that matterhorn will work as expected when installed
via the default cabal install mechanism instead of relying on more or
less hard coded non-standard paths which only exists for the custom
release builds.

This for example caused matterhorn to not work as expected in NixOS:
https://github.com/NixOS/nixpkgs/issues/129048