Matterhorn.Config: add syntax dir from getDataFileName to search path
Similarly to emoji.json, the syntax files are cabal data-files and are
installed via cabal as well and we can get the path to them via
getDataFileName, so we can add that directory to the default search path
as well.

Here we prefer the cabal data directory over the release directory
layout as well, just as with emoji.json.