Add support for mouse events in the /notify-prefs window
This change delegates incoming mouse events in the /notify-prefs window
to the form event handler for that window. Since the Brick library's
form event handler already supports mouse events to interact with the
form's fields, this change was mostly about refactoring our event
dispatching to forward mouse events to the form's handler.