Make keybinding consistency checker detect collisions within a mode
Previously, if the user remapped keys such that the same key was bound
to two different events in a single application mode, we would fail to
detect that and report it to the user. This change updates the collision
checker so that instead of building the full key handler maps by
resolving all key events into bindings and then checking the result (by
which time any collisions would have been silently ignored!), we build a
list of handlers for each mode and then check to see whether any of them
have duplicate entries. The code was trying to do this before, but we
didn't realize that collisions would result in the built-up key handler
maps eliminating collisions by construction. This change checks the
lists of handlers without building such maps, allowing us to check on
things before that map-building step has a chance to eliminate
collisions.