fetchVisibleIfNeeded: clear the cdFetchPending flag before calling addObtainedMessages
This change reorders these steps so that the cdFetchPending flag is
cleared first, because the previous implementation could conceivably
forget to clear the field if addObtainedMessages raised an exception for
any reason. Granted, that shouldn't happen because addObtainedMessages
shouldn't do IO -- it should queue up its IO asynchronously -- but this
ensures that such a state can never be reached regardless.