Use explicit visibility request from EventM, not render-time request, to show selected channel
Previously we used the 'visible' function from Brick to ensure that the
selected channel was visible in the scrollable channel list. Now we use
the functionally equivalent approach of calling 'makeVisible' from
EventM and 'reportExtent' at rendering time. But the difference is that
we have more control over when this visibility-request-triggered
scrolling occurs, which paves the way for gaining more control over the
scroll position of the channel list and decoupling it from the position
of the selected channel (which we may not always want to be visible if
we support scrolling the channel list).