setFocusWith: take an action to invoke only when a channel change occurs
This change also updates nextChannel so that if the channel selection
cycles from the end to the beginning of the channel list, the topmost
channel group header is scrolled into view. This improves what was
previously an awkard experience where once that heading scrolled out of
view, the only way to get it back was to manually scroll the channel
list up.