Fix last-run state saving per-team
This change fixes a bug where the current team's selected channel was
used to determine whether *all* last-run states got saved. That is bad
because it meant that if the current team's selected channel was a
normal one, other team states would get saved even if their selected
channels were DM channels. We never wanted to save in those cases, and
that led to startup crashes due to violating the assumption that all
initially-selected channels are non-DM ones.