Parameterize a lot of things on EditState
This change moves closer to a more self-contained editor abstraction
by parameterizing a lot of things on the editor to inspect/modify
rather than (in many, many places) assuming that the current team's
channel message editor is intended. This is an important incremental
move because we'd like to re-use a lot of the editor/autocomplete
functionality in a thread window, but to do that, we need to be able to
distinguish between more than one such editor per team: one for the main
channel UI, and one (or more, if we wanna get crazy down the road) for a
thread window.