Replace setMode with a mode stack
This change makes it possible to generalize mode management in many
places so that code that returns from one mode to the previous mode
needs to know less about the mode it might be returning to. This is done
by adding a mode "stack" so that returning to the previous mode is just
a matter of popping the current mode off of the stack. The
implementation prevents the stack from ever being empty; in practice
this means that Main will always be at the bottom of the stack, and
attempts to pop that mode will silently do nothing.

This change also removes some functionality from a few places where we
had specific cases to track previous modes that we needed to return to;
now that need is met by the mode stack in a more general way.