Parameterize message selection functionality on message list and selection state
This patch modifies the message selection functionality so that is
more decoupled from the channel message listing, editor, and selection
states. This makes it possible to use the message selection state
functions and event handlers on message selection UIs built elsewhere.

This change also modified the MessageSelect mode to carry the channel ID
associated with the selection.