Draw: draw modal modes by rendering each mode in the mode stack
This change makes individual mode renderers more composable by invoking
the mode renderer for each mode in a mode stack for all modes where we
know that we want to represent prior modes in lower layers. For modes
where we know that the main UI is all we need to draw, we just draw
that. Otherwise we invoke the draw function for every mode in the mode
stack. This makes modal renderings (such as for pop-up windows) much
more composable in new ways because they can be stacked on top of
arbitrary modes.

This also changes how monochrome rendering of the main UI is handled
by moving the monochrome logic out of drawMain and into the main mode
stack renderer. This makes it so that other layers now automatically get
rendered in monochrome if they're lower layers. This behavior can be
controlled in how the top-level draw routine renders lower layers on a
per-mode basis.