Get rid of ThreadWindowMessageSelect mode
This change moves the mode state of the thread window into ThreadWindow
itself as a new field of type ThreadInterfaceMode. This allows to have
one Mode for showing the window at all -- ThreadWindow -- and an
independently variable mode for drawing and event dispatching within the
window, making the window more composable. Before this change, the
additional Mode meant that window internals had to correspond to that
mode and we needed top-level dispatching cases for that mode that made
the ThreadInterface much more difficult to re-use.