Retool how event handler fallback works
Previously, the functions that got used to chain event handlers were
extremely clunky and difficult to compose, leading to a lot of very
verbose event handling fallback code. This change cleans up event
handler fallback significantly with the introduction of the
'handleEventWith' function, which takes a list of handlers to try and
tries them all, in order, until a handler handles an event.

This also removes the "fallback" argument from handleKeyboardEvent,
since previously the fallback logic was built into handleKeyboardEvent.
This change decouples the fallback logic from handleKeyboardEvent
itself, making for a cleaner and more composable API.