Undo tight coupling of spell check timers to teams
Previously, there was at most one spell check timer per team. While this
worked well enough, it was very inflexible to be refactored to use a
separate spell check timer for more than one editor interface per team,
which we need in an upcoming refactor. This change rearranges how spell
check timers are tracked so that we keep one around for each thread and
for each team editor, which is much easier to generalize to be
per-channel later on.