Monster patch to replace per-team EditState with per-channel EditState
This patch adds an EditState to each ClientChannel and uses that rather
than a single per-team EditState (formerly in the _tsEditState field of
TeamState).

This change is an incremental change toward making the threading
interface and the channel interface use the same message listing and
editor abstractions. Despite this patch touching many files, all it
does is make various spots use the EditState in ClientChannel. However,
that has some knock-on effects: the EditState now carries the TeamId (if
available) of its channel, its ChannelId, and many spots need those
values. Many functions that previously took a TeamId and assumed that
the current channel was the one of interest are now more generic and
look inside the EditState to get that information, making those
functions more contained.

This change *should* be functionality-preserving, but there are a few
spots where things got a bit tricky and needed to change (and may need
to change more as we discover issues). As an example, the autocomplete
implementation used to query the server to get commands for the current
team regardless of the channel type in which the autocomplete was
initiated. This is no longer true: because autocomplete triggers now
operate a specific EditState, if that EditState has no TeamId (which is
true for DM channels), then the autocomplete results will only contain
client commands because we can't ask the server for commands if we don't
have a TeamId. I think this is actually probably more correct despite
the fact that the web client does include server commands in the command
list when autocompleting in DM channels. If that ends up being an issue,
we can hack around it by assuming that the application's current team is
a fine source of TeamId for the purposes of making the API call.