Rename ThreadInterface type to MessageInterface and generalize
This change generalizes the ThreadInterface type so that it can be more
suitable for non-thread message interfaces (an incremental change toward
using this for channels, too). This change adds a requirement that the
interface be parameterized on whether it has a root post ID (for
threads) or not (for channels). That gives us a static type for ensuring
that we have the data when we need it. This change also renames the
fields and updates the comments to be more in line with this abstraction
being a bit more generic than it was previously. Lastly, this change
adds a "ThreadInterface" type alias that is used specifically for the
case where a post ID is provided when we mean "thread" interface rather
than "channel" interface (to be added in a future patch).