Make thread window display underneath main channel interface
This change modifies the UI so that the thread window is displayed
simultaneously with the main channel message interface. This change
brings Matterhorn closer to the behavior of the official web client and
also makes the thread UI more usable.

As part of this change, some new things were introduced:

 * A new key event, "change-message-editor-focus"
   (ChangeMessageEditorFocus), which defaults to M-o, for switching the
   input focus between the thread and channel editors

 * New theme attributes, "thread", "thread.header", and
   "focusedEditorPrompt", for styling the thread window area, the thread
   window's title bar, and the focused editor's prompt, respectively