Make thread window orientation configurable
This change adds a new configuration file setting, "threadOrientation",
that defaults to "below" and can take the values "above", "below",
"left", and "right". The options indicate the position of the thread
window relative to the main channel message area.

This change also removes threadTitleBarAttr (thread.header) since it is
no longer used.