openThreadWindow: get the state after the messages have been fetched
This ensures that the call to getMessageForPostId always succeeds since
the message it seeks will have just been added to the state. Previously
we got the state before the messages were added, which meant that it was
possible for getMessageForPostId to fail to find its message if it
happened to be newly fetched in this call to openThreadWindow.