Position the autocomplete layer relative to the editor prompt
This change fixes a problem where the autocomplete layer, which was
recently changed to position relative to the message editor, was being
positioned badly when the editor was moved due to the prompt changing
size, such as when a reply prompt was present. This change fixes this
problem by stabilizing the position of the autocomplete layer; it is now
positioned relative to the prompt coordinates, which will be stable
regardless of the prompt contents and thus the editor position.