Add configuration file setting to control how unread teams are sorted (#787)
This change adds a new setting, teamListSorting, that has the same value
choices as the channelListSorting setting. When set to "unread-first",
the setting causes unread teams to be floated to the beginning of the
team list after other ordering considerations have been applied (such as
user-configured team list ordering preferences).