Improve autocompletion by always attempting to select an initial match based on prefix
This change makes it so that when tab-completing to autocomplete a
user, channel, emoji, etc., then initial selection (when the first
tab-completion is requested) will be attempted based on a prefix string
match of the list of autocomplete results. This fixes a long-standing
behavior where, for example, autocompletion search for inputs starting
with 'f' would possibly yield results *containing* 'f' earlier in the
list than those starting with 'f' (such as those starting with 'a'), and
the first TAB keypress would complete to the first alternative starting
with 'a'. This strikes me as never what anyone wants, since the first
autocomplete result one is likely to want is the one that starts with
the input string even if there are also alternatives that contain it.