Use existing post reaction and file metadata rather than making additional API requests to fetch it
This change drastically reduces the number of API requests that
matterhorn makes when obtaining posts. Previously, it made O(n) requests
to get reaction and file attachment metadata. We didn't know that the
data was already provided in Post structures in the API response (and we
don't know when that was added to the API). But using that, as is now
done in this change, greatly improves the application's responsiveness,
at startup and on channel (re)load, even on fast connections.