Support new "auto" mode to determine channel list width
This change adds a new 'auto' value for the 'channelListWidth'
configuration setting. Its default is unchanged and it can still take an
integer value specifying the channel list width in columns. If set to
'auto', its value is determined by Matterhorn, bounded by minimum and
maximum values, as a function of window width. This is intended to give
nicer results as font size decreases, up to a point.