Ignore DM channels' mentions when computing unread status
This change was made because without it, if you are viewing a DM
channel, then the other user's posts in that channel would cause that DM
channel to briefly show up in the channel list as unread and then read
again, while we were in between getting the post and waiting for the
server's updated-viewed websocket response. This is because hasUnread'
was checking the channel's mention count. But it turns out that DM
channel messages always implicitly count as a mention, which leads to
the UI behavior described here. We never want that and don't really
care about this overloading of the mention count, so this change makes
it so that a channel is only unread by mention count if it is not a
Group or Direct channel.