Update for Vty 6 and enable cross-platform builds
This mostly-functionality-preserving change updates Matterhorn to build
with Vty 6 and vty-crossplatform, meaning that this change ought, at
least in theory, to enable Windows support for Matterhorn.

This change does include one known regression, though: it removes
support for the 'stty erase' character mapping discussed and fixed here:

https://github.com/matterhorn-chat/matterhorn/issues/565

This is because these changes move Matterhorn away from accessing
Unix-specific internals. I'm not exactly sure where the old logic should
live, though. On one hand, it could just go in vty-unix and be an
implicit part of how that backend sets up Vty. On the other hand, I'm
not entirely sold on the idea that 'stty erase' should always be implicitly
honored and it strikes me more as an application policy issue. (So by
that reasoning the logic might be implemented by vty-unix, but it should
not be on by default.) In any case, that regression should be resolved
one way or another before cross-platform support in Matterhorn is
considered ready for release.