preChangeChannelCommon: clear edit threshold
This change updates the channel change process to clear the edit
threshold of the channel that we're switching away from. This is aimed
at ensuring that the sidebar doesn't show the channel in question as
having unread activity, since it would do so if there is an edit that
hasn't been viewed. Since it's the channel the user was already viewing,
we want to reset the edit threshold so that it doesn't appear as unread.