openWithOpener: run the provided fetcher in the async worker
Prior to this change, if openWithOpener was given a fetcher that took
lots of time (e.g. due to fetching a very large file), it would run that
fetcher itself, causing it to block. This caused the main event loop to
stall while the fetch was going on; for small files this was never
noticeable, but for very large ones the effect was that the program
would appear to hang (even though it would eventually continue handling
events).

This change makes it so that the initial fetch step is done in the async
worker thread, allowing input events and server messages to continue to
be handled in the mean time.