checkForAutocompletion: only call with a traversal that can fail to look up (fixes #832)
Prior to this change, code that invoked the autocompletion routines gave
it a traversal that was usually used as a lens and would raise an
exception if the traversal didn't find a target. That was because in
almost all cases where we want to find a message interface by using a
lens, it's done synchronously and we expect it to exist in the
application state, so to avoid dealing with a traversal that should
never fail to find its target, we make the traversal unforgiving by
using 'singular' in it.

That was a problem for autocomplete code because it was specifically
designed to cope with the situation where the results of an
autocompletion query might apply to a channel that the user was removed
from since the autocompletion query was scheduled in the async thread,
meaning that when the autocompletion results come in we need to be ready
to discard them if their target channel is no longer in the application
state.

However, since we were giving the autocompletion code the unforgiving
version of the traversal to locate the channel structure, that would
result in async work being queued up that would ultimately raise an
exception rather than cope with the missing channel. Unfortunately the
problem wasn't detectable with static types because we were providing a
lens, which is compatible with the traversal requirement of the
autocomplete code. That meant that the problem was that the call sites
for checkForAutocompletion needed to be changed to provide a forgiving
version of the same traversal they were already getting so as to get the
benefit of how the autocompletion code is already designed to handle
this situation.

This was discovered in a situation where typing "/leave" and leaving a
channel before an autocomplete response for commands was received would
cause the autocompletion response handler to run the unforgiving
traversal to update the channel state for the channel that was being
left, triggering a runtime exception.