Autocomplete: fix support for completion in thread message interfaces
A previous recent commit resolved an issue with async work for
autocompletion, but in the process it broke autocompletion in threads
because the fix only ever worked for message interfaces for channels.
This fixes that problem, but to make it work, we needed a way to figure
out which kind of message editor to use in the autocomplete logic
without using prebuilt traversals that could fail (see prior related
change).

This works by introducing a new EditState field of type EditorTarget,
which is analogous to the MessageInterfaceTarget type. (We don't re-use
that type because doing so would require renaming it and moving it into
EditState, and then its name would be the wrong name for its use in
MessageInterface.) The EditState carries around an EditorTarget, which
is then enough information to use to figure out at autocomplete time
which editor to manipulate from the application state, and thus which
traversal to build to locate it.