writeBChan: ignore failed BChan writes
Prior to this change, a full application event channel would cause an
exception and a crash when an event write occurred. That is very rare,
but in particular, it isn't something we can do anything about and it
also isn't something the user can do anything about. The best we can do
is ignore the failed write on the assumption that the main event loop
will make room in the event channel for later writes. A downside risk of
doing this is that the application will miss events that would have led
to important state updates. But that's a small risk given that the only
circumstance where a full channel has been reported has been under
pathological network conditions.