Fix the -i/--ignore-configuration option. (Fixes #847)
Make the code that loads the configuration explicitly return the default
configuration without any config file lookups if the option to ignore
the configuration is used. The current implementation tries to make the
function that looks up the configuration file not load anything by
passing a null value to it, but that behavior clashes with the behavior
of the `-c`/`--config` option, which when not used results in the same
value being passed to the function, so it doesn't have a way of
differentiating between the two. As a result, if you have
a configuration file in any of the default locations it'll load it
regardless.