Add the `-i` option to a command in the release build script
The step that checks if the command reference Markdown document is
current is not using the `-i` option to ignore the configuration. That
shouldn't be a problem on almost any case because the configuration
doesn't affect the list of commands in any way, but if there's something
in the user's configuration that triggers a config validation error (for
example, in the case that one is using a config option implemented in
one branch but not yet included in the release) then the build doesn't
succeed.