Themes: use 'digest' package's CRC32 to hash usernames
Prior to this change, we used the 'hashable' package's hash to map
usernames to username colors. It turns out that such hashes are
unstable, which led to username colors changing occasionally (such as
from one release to the next). To fix this, we needed to use a stable
hash to choose the username buckets, which this commit does by using the
CRC32 hash from the 'digest' package instead of 'hashable'.