channelInfoFromChannelWithData: preserve channel's URL name in ChannelInfo
Prior to this change, both cdName and cdDisplayName fields of
ChannelInfo were populated with a channel's display name. That led to
/focus not working at all when its argument needed to be matched on
channels by URL name, which is actually the default behavior due to the
fact that tab completion completes input to channel names that should be
matched in this way.

This change ensures that the channel's URL name is preserved in cdName
separately from its display name in cdDisplayName.