Autocompletion: fix a race condition
This fixes a race where tab completion could complete the wrong kind of
alternative based on the input. This would happen when:

* User enters some kind of completable text of type A (e.g. slash
  command)
* User tab-completes an alternative from the list of A alternatives
* User enters some new kind of completable text of type B (e.g. @ user
  mention)
* User tab-completes a user alternative, but gets a slash command
  inserted instead

This change updates the autocompletion logic to record the kind of
completion that is in flight based on type; previously we tracked the
string that prompted the completion search, but that's not enough. By
tracking the type as well, and by clearing the autocomplete list
whenever a new kind of completion text is entered (but before the server
response is received), we can be sure to only insert a completion
alernative when the list of alternatives present matches the type of the
completed text.