editPostInOpenThread: propagate edits for thread root posts in addition to non-root posts
Prior to this change, edits to the root post of a thread wouldn't get
propagated to that post in a thread view for that thread. This happened
because we only ever checked the root post ID for the edited post to see
if it was in the open thread, but the root post itself doesn't have its
own root post ID set at all. This change fixes the behavior by checking
that the edited post is either in the thread (via the root post ID) or
is itself the root.